* origin/master: (63 commits)
  das: let ELLE_DAS_CLI_SYMBOL's be non positional by default and no short option
  crash-report: disable the test
  elle: update
  cli.Silo: simplify
  cli.Device: simplify
  cli.Doctor: simplify
  memo.environ: don't accept getenv on unknown names
  memo: clean http support
  cli.Network: Add create_mutable operation to script mode.
  beyond: fix the debug-symbol checkout
  drakefile: simplify logic
  memo.kvs: Fix dockerized python client.
  CI: fix
  CI: prepare for build from scratch
  CI: generate the symbols before checking
  memo: use elle::print
  docker: clean up alpine
  CI: fix address of the package host
  memo.kelips: fix variable name
  ChangeLog: 0.9.2
  ...

+++ beyond/src/__init__.py
@@ -58,12 +58,13 @@ def symbols_update():
   if symbols_checkout():
     return True
   # We failed.  Maybe the repository is broken, erase it and restart.
-  try:
-    shutil.rmtree(repo_dir)
-  except Exception as e:
-    log('symbolize: symbols_update: cannot remove {}: {}',
-        repo_dir, e)
-    return False
+  if os.path.exists(repo_dir):
+    try:
+      shutil.rmtree(repo_dir)
+    except Exception as e:
+      log('symbolize: symbols_update: cannot remove {}: {}',
+          repo_dir, e)
+      return False
   p = run(['git', 'clone',
            'git@git.infinit.sh:infinit/debug-symbols.git',
            repo_dir])
@@ -112,7 +113,7 @@ os.environ['MEMO_CRASH_REPORT'] = '0'
 
 def find_binaries():
   for path in chain(
-      [os.environ.get('MEMO_BINARIES')],
+      [os.environ.get('BIN_DIR')],
       os.environ.get('PATH', '').split(':'),
       ['bin', '/opt/memo/bin'],
   ):
